ARM: dts: rockchip: Update LEDs to work with the last leds-gpio driver on radxarock
The leds-gpio driver recently switched to the device property API. The device_node
name is no longer retrieved if the "label" devicetree property is not found.
In this case the driver tries to create entries with (null) name in
/sys/class/leds, which is wrong and generates backtrace as several gpio_leds have
the same name. Also renamed subnode "yellow" to "blue" to match the last
schematics updates.

Signed-off-by: Romain Perier <romain.perier@gmail.com>