usb: dwc2: Register interrupt handler only once gadget is correctly initialized
Don't register interrupt handler becore usb gadget is correctly initialized.
For embedded platforms which don't have a usb-phy it will crash the driver
because an interrupt might be emitted with non-initiliazed hardware.
According to devm_request_irq documentation, an interrupt can be emitted
at any time once the interrupt is registered, so we have to care about driver
and hardware initialization.

Signed-off-by: Romain Perier <romain.perier@gmail.com>